requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
python-multipart
